Add Level 500 and hero and SPLEvel

+++ Kint-Sources-master/OpenNos.Data/CharacterDTO.cs
@@ -68,7 +68,7 @@ namespace OpenNos.Data
 
         public bool HeroChatBlocked { get; set; }
 
-        public byte HeroLevel { get; set; }
+        public int HeroLevel { get; set; }
 
         public long HeroXp { get; set; }
 
@@ -82,7 +82,7 @@ namespace OpenNos.Data
 
         public long LastFamilyLeave { get; set; }
 
-        public byte Level { get; set; }
+        public int Level { get; set; }
 
         public long LevelXp { get; set; }
